fix bug in worker
